+++ examples/app_text_elements/app.tsx
@@ -10,22 +10,17 @@ import {
   TextInput,
   Title,
 } from "@canva/app-ui-kit";
-import { initAppElement } from "@canva/design";
+import { initAppElement, FontWeight, TextAttributes } from "@canva/design";
 import React from "react";
 import styles from "styles/components.css";
-
-type FontWeight = "normal" | "bold";
-type FontStyle = "normal" | "italic";
-type Decoration = "none" | "underline";
-type TextAlign = "start" | "center" | "end";
 
 type AppElementData = {
   text: string;
   color: string;
   fontWeight: FontWeight;
-  fontStyle: FontStyle;
-  decoration: Decoration;
-  textAlign: TextAlign;
+  fontStyle: TextAttributes["fontStyle"];
+  decoration: TextAttributes["decoration"];
+  textAlign: TextAttributes["textAlign"];
   width: number;
   rotation: number;
   useCustomWidth: boolean;
@@ -129,7 +124,7 @@ export const App = () => {
           label="Font style"
           value={fontStyle}
           control={(props) => (
-            <Select<FontStyle>
+            <Select<TextAttributes["fontStyle"]>
               {...props}
               options={[
                 { value: "normal", label: "Normal" },
@@ -155,7 +150,13 @@ export const App = () => {
               {...props}
               options={[
                 { value: "normal", label: "Normal" },
+                { value: "thin", label: "Thin" },
+                { value: "extralight", label: "Extra light" },
+                { value: "light", label: "Light" },
+                { value: "medium", label: "Medium" },
+                { value: "semibold", label: "Semibold" },
                 { value: "bold", label: "Bold" },
+                { value: "heavy", label: "Heavy" },
               ]}
               onChange={(value) => {
                 setState((prevState) => {
@@ -173,7 +174,7 @@ export const App = () => {
           label="Decoration"
           value={decoration}
           control={(props) => (
-            <Select<Decoration>
+            <Select<TextAttributes["decoration"]>
               {...props}
               options={[
                 { value: "none", label: "None" },
@@ -195,7 +196,7 @@ export const App = () => {
           label="Text align"
           value={textAlign}
           control={(props) => (
-            <Select<TextAlign>
+            <Select<TextAttributes["textAlign"]>
               {...props}
               options={[
                 { value: "start", label: "Start" },